scripts/requirements.txt
requests>=2.31.0
supabase==2.4.6
python-dotenv>=1.0.1
beautifulsoup4>=4.12.0
PyMuPDF>=1.24.0
